pkg: move packages from cni to plugins
