--- pkg/ns/ns_suite_test.go
@@ -1,34 +1,0 @@
-// Copyright 2016 CNI authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ns_test
-
-import (
-	"math/rand"
-	"runtime"
-
-	. "github.com/onsi/ginkgo"
-	"github.com/onsi/ginkgo/config"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestNs(t *testing.T) {
-	rand.Seed(config.GinkgoConfig.RandomSeed)
-	runtime.LockOSThread()
-
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "pkg/ns Suite")
-}
